% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Mean of numeric vector}
\usage{
mymean(numbers)
}
\arguments{
\item{numbers}{A vector of numbers as an input.}
}
\value{
A number which represents the average of the input vector.
}
\description{
\code{mymean()} calculates the mean or average value of a given numeric vector and outputs this mean value.
}
\examples{
mymean(c(1, 6, 8, 1))
mymean(c(10, 5, 3, 11, 9))
}
